+++ src/FluentAssertions.Optional.Tests/Numeric/OptionalNumericAssertionsTests.cs
@@ -3,7 +3,7 @@ using Optional;
 using Xunit;
 using Xunit.Sdk;
 
-namespace FluentAssertions.Optional.Tests
+namespace FluentAssertions.Optional.Tests.Numeric
 {
     public class OptionalNumericAssertionsTests
     {
